requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Reinforcement Learning
gymnasium>=0.28.0
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0

# Web Interface
gradio>=4.0.0

# Data Processing and Utilities
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Game Theory and Math
scipy>=1.7.0
networkx>=2.6.0

# Model Management
huggingface-hub>=0.16.0
datasets>=2.10.0

# Testing and Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.10.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.28.0 